Revert "MAN-2067 fix dependency cycle build errors"
This reverts commit fa693ef1c8bfdcaa00a30c86a7da4adeb2417c2b.
