+++ server/@types/express/index.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-cycle */
 import type { UserDetails } from '../../services/userService'
 import type { ActivityLogCache } from '../index'
 import { Errors } from '../../models/Errors'
