+++ server/@types/index.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/no-cycle
 export * from './Route.type'
 export * from './Controller.type'
 export * from './FileUpload.type'
