+++ server/data/index.ts
@@ -13,6 +13,7 @@ const applicationInfo = applicationInfoSupplier()
 initialiseAppInsights()
 buildAppInsightsClient(applicationInfo.applicationName)
 
+// eslint-disable-next-line import/no-cycle
 import HmppsAuthClient from './hmppsAuthClient'
 import ManageUsersApiClient from './manageUsersApiClient'
 import { createRedisClient } from './redisClient'
