+++ server/data/manageUsersApiClient.ts
@@ -1,5 +1,6 @@
 import logger from '../../logger'
 import config from '../config'
+// eslint-disable-next-line import/no-cycle
 import RestClient from './restClient'
 
 export interface User {
