+++ server/data/restClient.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-cycle */
 import { HttpAgent as Agent, HttpsAgent } from 'agentkeepalive'
 import superagent, { Response } from 'superagent'
 import logger from '../../logger'
@@ -5,9 +6,7 @@ import sanitiseError from '../sanitisedError'
 import type { ApiConfig } from '../config'
 import { restClientMetricsMiddleware } from './restClientMetricsMiddleware'
 import { ErrorSummaryItem } from './model/common'
-import { escapeForLog } from '../utils/escapeForLog'
-import { isValidHost } from '../utils/isValidHost'
-import { isValidPath } from '../utils/isValidPath'
+import { escapeForLog, isValidHost, isValidPath } from '../utils'
 import 'multer'
 
 interface Request {
