+++ server/models/Locals.ts
@@ -1,5 +1,6 @@
-import { RiskData } from '@ministryofjustice/hmpps-arns-frontend-components-lib'
+/* eslint-disable import/no-cycle */
 import { Response } from 'express'
+import { RiskData } from '@ministryofjustice/hmpps-arns-frontend-components-lib'
 import { PersonalDetails } from '../data/model/personalDetails'
 import { FeatureFlags } from '../data/model/featureFlags'
 import { Sentence } from '../data/model/sentenceDetails'
@@ -20,11 +21,12 @@ import { PersonRiskFlags, RiskScore, RiskSummary, RoshRiskWidgetDto, TimelineIte
 import { TierCalculation } from '../data/tierApiClient'
 import { ErrorSummary } from '../data/model/common'
 import { Activity, PersonAppointment, PersonSchedule } from '../data/model/schedule'
-import { FileCache } from '../@types/FileUpload.type'
+import { FileCache } from '../@types'
 import { SentencePlan } from './Risk'
 import { ContactResponse } from '../data/model/overdueOutcomes'
 import { SmsPreviewResponse } from '../data/model/OutlookEvent'
 import { ESupervisionCheckIn, OffenderCheckinsByCRNResponse } from '../data/model/esupervision'
+import { outcomeOptions } from '../properties'
 
 export interface AppointmentLocals {
   meta: {
