+++ server/utils/decorateFormAttributes.ts
@@ -2,6 +2,7 @@
 import { Request } from 'express'
 import { getDataValue } from './getDataValue'
 import config from '../config'
+// eslint-disable-next-line import/no-cycle
 import { AppResponse } from '../models/Locals'
 
 interface Item {
