+++ server/utils/roleDescription.test.ts
@@ -1,4 +1,4 @@
-import { Contact } from '../data/model/personalDetails'
+import { Contact } from '../data/model/professionalContact'
 import { roleDescription } from './roleDescription'
 
 const mockContact = {
